requirements.txt
# Core
fastapi
uvicorn
streamlit
python-dotenv

# Database (MySQL + SQLite)
sqlalchemy
pymysql[rsa]   # MySQL driver
# sqlite3 is built into Python, no need to list separately

# Ollama Integration
ollama   # Python client for Ollama

# NLP / Resume Parsing
spacy
pdfminer.six   # for PDF parsing
python-docx    # for DOCX parsing

# Utilities
pandas
numpy
requests
matplotlib
pdfminer

# Testing
pytest
